# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: The Waffle Hub BCM - KOT & Billing API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders with optional filters
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrder
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    delete:
      operationId: deleteOrder
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/items:
    post:
      operationId: addOrderItem
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderItem"
  /orders/{id}/items/{itemId}:
    patch:
      operationId: updateOrderItem
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderItemUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderItem"
    delete:
      operationId: deleteOrderItem
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /orders/{id}/payment:
    get:
      operationId: getOrderPayment
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
    post:
      operationId: createOrderPayment
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
    patch:
      operationId: updateOrderPayment
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # ── Customers ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Order:
      type: object
      required: [id, orderNumber, customerName, orderType, status, totalAmount, createdAt, readyTime, items]
      properties:
        id:           { type: integer }
        orderNumber:  { type: string }
        customerId:   { type: ["integer", "null"] }
        customerName: { type: string }
        customerPhone: { type: ["string", "null"] }
        orderType:    { type: string }
        status:       { type: string }
        notes:        { type: ["string", "null"] }
        totalAmount:  { type: number }
        createdAt:    { type: string }
        readyTime:    { type: string }
        updatedAt:    { type: string }
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        payment:
          $ref: "#/components/schemas/Payment"
    OrderInput:
      type: object
      required: [customerName, orderType]
      properties:
        customerName:  { type: string, minLength: 1 }
        customerPhone: { type: string }
        customerId:    { type: ["integer", "null"] }
        orderType:     { type: string }
        notes:         { type: string }
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
    OrderUpdate:
      type: object
      properties:
        customerName:  { type: string }
        customerPhone: { type: string }
        orderType:     { type: string }
        notes:         { type: ["string", "null"] }
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string }

    # ── Order Item ────────────────────────────────────────────────────────────
    OrderItem:
      type: object
      required: [id, orderId, productName, price, quantity]
      properties:
        id:          { type: integer }
        orderId:     { type: integer }
        productId:   { type: ["integer", "null"] }
        productName: { type: string }
        price:       { type: number }
        quantity:    { type: integer }
        notes:       { type: ["string", "null"] }
    OrderItemInput:
      type: object
      required: [productName, price, quantity]
      properties:
        productId:   { type: ["integer", "null"] }
        productName: { type: string, minLength: 1 }
        price:       { type: number }
        quantity:    { type: integer, minimum: 1 }
        notes:       { type: string }
    OrderItemUpdate:
      type: object
      properties:
        quantity: { type: integer, minimum: 1 }
        notes:    { type: ["string", "null"] }

    # ── Payment ───────────────────────────────────────────────────────────────
    Payment:
      type: object
      required: [id, orderId, totalAmount, cashAmount, upiAmount, cardAmount, totalPaid, status, createdAt]
      properties:
        id:          { type: integer }
        orderId:     { type: integer }
        totalAmount: { type: number }
        cashAmount:  { type: number }
        upiAmount:   { type: number }
        cardAmount:  { type: number }
        totalPaid:   { type: number }
        balance:     { type: number }
        status:      { type: string }
        createdAt:   { type: string }
    PaymentInput:
      type: object
      required: [totalAmount]
      properties:
        totalAmount: { type: number }
        cashAmount:  { type: number }
        upiAmount:   { type: number }
        cardAmount:  { type: number }
    PaymentUpdate:
      type: object
      properties:
        cashAmount: { type: number }
        upiAmount:  { type: number }
        cardAmount: { type: number }

    # ── Customer ──────────────────────────────────────────────────────────────
...