# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * The Waffle Hub BCM - KOT & Billing API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { Payment } from './payment';

export interface Order {
  id: number;
  orderNumber: string;
  /** @nullable */
  customerId?: number | null;
  customerName: string;
  /** @nullable */
  customerPhone?: string | null;
  orderType: string;
  status: string;
  /** @nullable */
  notes?: string | null;
  totalAmount: number;
  createdAt: string;
  readyTime: string;
  updatedAt?: string;
  items: OrderItem[];
  payment?: Payment;
}